// src/models/Users.ts
export interface Users {
  _id: string;
  ref: string;
  username: string;
  status: string;
  organization: string;
  email: string;
  bvn: string;
  date_joined: string;
  marital_status: string;
  children: number;
  gender: string;
  phone: string;
  type_of_residence: string;
  level_of_education: string;
  employment_status: string;
  sector_of_employment: string;
  duration_of_employment: string;
  office_email: string;
  monthly_income: string;
  loan_repayment: number;
  twitter: string;
  facebook: string;
  instagram: string;
  guarantor_full_name: string;
  guarantor_phone: string;
  guarantor_email: string;
  account_number: string;
  bank: string;
  guarantor_relationship: string;
}